requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
Django==1.10.6
django-crispy-forms==1.6.1
django-formtools==2.0
django-pagedown==0.1.3
django-reversion==2.0.8
future==0.16.0
httplib2==0.10.3
idna==2.5
Markdown==2.6.8
packaging==16.8
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
